Rename goes_nc -> nc_goes
