+++ satpy/etc/readers/nc_goes.yaml
@@ -4,42 +4,42 @@
 
 reader:
   description: GOES netCDF Reader
-  name: goes_nc
+  name: nc_goes
   sensors: [goes_imager]
   default_channels: []
   reader: !!python/name:satpy.readers.yaml_reader.FileYAMLReader
 
 
 file_types:
-    goes_nc_00_7:
-        file_reader: !!python/name:satpy.readers.goes_nc.GOESNCFileHandler
+    nc_goes_00_7:
+        file_reader: !!python/name:satpy.readers.nc_goes.GOESNCFileHandler
         file_patterns: ['{platform}.{start_time:%Y.%j.%H%M%S}.BAND_01.nc']
-    goes_nc_03_9:
-        file_reader: !!python/name:satpy.readers.goes_nc.GOESNCFileHandler
+    nc_goes_03_9:
+        file_reader: !!python/name:satpy.readers.nc_goes.GOESNCFileHandler
         file_patterns: ['{platform}.{start_time:%Y.%j.%H%M%S}.BAND_02.nc']
-    goes_nc_06_5:
-        file_reader: !!python/name:satpy.readers.goes_nc.GOESNCFileHandler
+    nc_goes_06_5:
+        file_reader: !!python/name:satpy.readers.nc_goes.GOESNCFileHandler
         file_patterns: ['goes12.{start_time:%Y.%j.%H%M%S}.BAND_03.nc',
                         'goes13.{start_time:%Y.%j.%H%M%S}.BAND_03.nc',
                         'goes14.{start_time:%Y.%j.%H%M%S}.BAND_03.nc',
                         'goes15.{start_time:%Y.%j.%H%M%S}.BAND_03.nc']
-    goes_nc_06_8:
-        file_reader: !!python/name:satpy.readers.goes_nc.GOESNCFileHandler
+    nc_goes_06_8:
+        file_reader: !!python/name:satpy.readers.nc_goes.GOESNCFileHandler
         file_patterns: ['goes08.{start_time:%Y.%j.%H%M%S}.BAND_03.nc',
                         'goes09.{start_time:%Y.%j.%H%M%S}.BAND_03.nc',
                         'goes10.{start_time:%Y.%j.%H%M%S}.BAND_03.nc',
                         'goes11.{start_time:%Y.%j.%H%M%S}.BAND_03.nc']
-    goes_nc_10_7:
-        file_reader: !!python/name:satpy.readers.goes_nc.GOESNCFileHandler
+    nc_goes_10_7:
+        file_reader: !!python/name:satpy.readers.nc_goes.GOESNCFileHandler
         file_patterns: ['{platform}.{start_time:%Y.%j.%H%M%S}.BAND_04.nc']
-    goes_nc_12_0:
-        file_reader: !!python/name:satpy.readers.goes_nc.GOESNCFileHandler
+    nc_goes_12_0:
+        file_reader: !!python/name:satpy.readers.nc_goes.GOESNCFileHandler
         file_patterns: ['goes08.{start_time:%Y.%j.%H%M%S}.BAND_05.nc',
                         'goes09.{start_time:%Y.%j.%H%M%S}.BAND_05.nc',
                         'goes10.{start_time:%Y.%j.%H%M%S}.BAND_05.nc',
                         'goes11.{start_time:%Y.%j.%H%M%S}.BAND_05.nc']
-    goes_nc_13_3:
-        file_reader: !!python/name:satpy.readers.goes_nc.GOESNCFileHandler
+    nc_goes_13_3:
+        file_reader: !!python/name:satpy.readers.nc_goes.GOESNCFileHandler
         file_patterns: ['goes12.{start_time:%Y.%j.%H%M%S}.BAND_06.nc',
                         'goes13.{start_time:%Y.%j.%H%M%S}.BAND_06.nc',
                         'goes14.{start_time:%Y.%j.%H%M%S}.BAND_06.nc',
@@ -62,7 +62,7 @@ datasets:
     coordinates:
         - longitude_00_7
         - latitude_00_7
-    file_type: goes_nc_00_7
+    file_type: nc_goes_00_7
 
   '03_9':
     name: '03_9'
@@ -80,7 +80,7 @@ datasets:
     coordinates:
         - longitude_03_9
         - latitude_03_9
-    file_type: goes_nc_03_9
+    file_type: nc_goes_03_9
 
   '06_5':
     name: '06_5'
@@ -98,7 +98,7 @@ datasets:
     coordinates:
         - longitude_06_5
         - latitude_06_5
-    file_type: goes_nc_06_5
+    file_type: nc_goes_06_5
 
   '06_8':
     name: '06_8'
@@ -116,7 +116,7 @@ datasets:
     coordinates:
         - longitude_06_8
         - latitude_06_8
-    file_type: goes_nc_06_8
+    file_type: nc_goes_06_8
 
   '10_7':
     name: '10_7'
@@ -134,7 +134,7 @@ datasets:
     coordinates:
         - longitude_10_7
         - latitude_10_7
-    file_type: goes_nc_10_7
+    file_type: nc_goes_10_7
 
   '12_0':
     name: '12_0'
@@ -152,7 +152,7 @@ datasets:
     coordinates:
         - longitude_12_0
         - latitude_12_0
-    file_type: goes_nc_12_0
+    file_type: nc_goes_12_0
 
   '13_3':
     name: '13_3'
@@ -170,88 +170,88 @@ datasets:
     coordinates:
         - longitude_13_3
         - latitude_13_3
-    file_type: goes_nc_13_3
+    file_type: nc_goes_13_3
 
   longitude_00_7:
       name: longitude_00_7
-      file_type: goes_nc_00_7
+      file_type: nc_goes_00_7
       standard_name: longitude
       units: degrees_east
 
   latitude_00_7:
     name: latitude_00_7
-    file_type: goes_nc_00_7
+    file_type: nc_goes_00_7
     standard_name: latitude
     units: degrees_north
 
   longitude_03_9:
       name: longitude_03_9
-      file_type: goes_nc_03_9
+      file_type: nc_goes_03_9
       standard_name: longitude
       units: degrees_east
 
   latitude_03_9:
     name: latitude_03_9
-    file_type: goes_nc_03_9
+    file_type: nc_goes_03_9
     standard_name: latitude
     units: degrees_north
 
   longitude_06_5:
       name: longitude_06_5
-      file_type: goes_nc_06_5
+      file_type: nc_goes_06_5
       standard_name: longitude
       units: degrees_east
 
   latitude_06_5:
     name: latitude_06_5
-    file_type: goes_nc_06_5
+    file_type: nc_goes_06_5
     standard_name: latitude
     units: degrees_north
 
   longitude_06_8:
       name: longitude_06_8
-      file_type: goes_nc_06_8
+      file_type: nc_goes_06_8
       standard_name: longitude
       units: degrees_east
 
   latitude_06_8:
     name: latitude_06_8
-    file_type: goes_nc_06_8
+    file_type: nc_goes_06_8
     standard_name: latitude
     units: degrees_north
 
   longitude_10_7:
       name: longitude_10_7
-      file_type: goes_nc_10_7
+      file_type: nc_goes_10_7
       standard_name: longitude
       units: degrees_east
 
   latitude_10_7:
     name: latitude_10_7
-    file_type: goes_nc_10_7
+    file_type: nc_goes_10_7
     standard_name: latitude
     units: degrees_north
 
   longitude_12_0:
       name: longitude_12_0
-      file_type: goes_nc_12_0
+      file_type: nc_goes_12_0
       standard_name: longitude
       units: degrees_east
 
   latitude_12_0:
     name: latitude_12_0
-    file_type: goes_nc_12_0
+    file_type: nc_goes_12_0
     standard_name: latitude
     units: degrees_north
 
   longitude_13_3:
       name: longitude_13_3
-      file_type: goes_nc_13_3
+      file_type: nc_goes_13_3
       standard_name: longitude
       units: degrees_east
 
   latitude_13_3:
     name: latitude_13_3
-    file_type: goes_nc_13_3
+    file_type: nc_goes_13_3
     standard_name: latitude
     units: degrees_north
