+++ satpy/readers/nc_goes.py
