+++ satpy/tests/reader_tests/__init__.py
@@ -34,7 +34,7 @@ from satpy.tests.reader_tests import (test_abi_l1b, test_hrit_base,
                                       test_seviri_calibration, test_clavrx,
                                       test_grib, test_hrit_goes, test_ahi_hsd,
                                       test_iasi_l2, test_generic_image,
-                                      test_scmi, test_hrit_jma, test_goes_nc)
+                                      test_scmi, test_hrit_jma, test_nc_goes)
 
 if sys.version_info < (2, 7):
     import unittest2 as unittest
@@ -69,6 +69,6 @@ def suite():
     mysuite.addTests(test_generic_image.suite())
     mysuite.addTests(test_scmi.suite())
     mysuite.addTests(test_hrit_jma.suite())
-    mysuite.addTests(test_goes_nc.suite())
+    mysuite.addTests(test_nc_goes.suite())
 
     return mysuite
